fix: pre-audit security hardening (10 findings) (#355)
1. Add sanitize_fts_query() to RRF hybrid search path (was missing defense-in-depth)
2. Reject missing HNSW checksum files instead of warning (prevents bincode deser of unverified data)
3. Containment check on ref remove — canonicalize + starts_with before remove_dir_all
4. Reject symlink reference paths instead of warning (trust boundary)
5. 10MB size guard on display file reads (was unbounded)
6. Size guards on config (1MB), notes (10MB), project registry (1MB) reads
7. saturating_add(1) in similar command (overflow on usize::MAX)
8. PID-based temp file suffix for notes (was predictable docs/notes.toml.tmp)
9. 500K edge cap on call graph query (was unbounded SELECT)
10. Clamp trace --depth to 1-50 via clap value_parser

+++ src/cli/mod.rs
@@ -220,9 +220,9 @@ enum Commands {
         source: String,
         /// Target function name or file:function
         target: String,
-        /// Max search depth
-        #[arg(long, default_value = "10")]
-        max_depth: usize,
+        /// Max search depth (1-50)
+        #[arg(long, default_value = "10", value_parser = clap::value_parser!(u16).range(1..=50))]
+        max_depth: u16,
         /// Output format: text, json, mermaid
         #[arg(long, default_value = "text")]
         format: String,
@@ -375,7 +375,7 @@ pub fn run_with(mut cli: Cli) -> Result<()> {
             json,
         }) => {
             let fmt = if json { "json" } else { format.as_str() };
-            cmd_trace(&cli, source, target, max_depth, fmt)
+            cmd_trace(&cli, source, target, max_depth as usize, fmt)
         }
         Some(Commands::TestMap {
             ref name,
